:package: add a PHP version of the foundation index page and a header template
This is the beginning of the PHP template system for the Foundation
front-end framework. This is going to replace the Bootstrap
templates with the Foundation templates.

This commit kick-starts things by adding a PHP file with the
contents of the HTML version of the foundation index page. It
also adds the first foundation template, which is for the header.