:package: add PHP version of new 'Projects' page
Adds a PHP version of the new Projects page, which has been
redesigned using the Foundation framework.
Note that an undefined variable is used in the hyperlink for
MineCart. This will be fixed later when it's added to the
config, but for now this will produce a pretty useless link.