:wrench: delete Foundation HTML pages
Removes the HTML versions of the re-designed
pages that use the Foundation framework.
All content was checked in the new PHP
versions to ensure that the PHP pages
mirrored the old HTML pages, so no content
has been lost. There are now PHP versions of
all Foundation pages, except for the SmartPhix
blog posts.