:bug: update requires and links in all blog posts
Fixes the links in the blog posts and the SmartPhix
tag index page. They now use the correct renamed
template files. The SmartPhix tag link was also
updated to have "index.php" at the end, so that
it shows up in a "Find All" search for ".php".
This will hopefully make it easier to spot internal
links that need updating in the future.

This commit also fixes links in the "Welcome to
Redhawk" blog post that point to the old projects
and contact pages.