:sparkles: update copyright message in footer
A very old README file pointed out that the project
start date was 25/10/2012. The start year for the
copyright message was updated accordingly, changing
from 2014 to 2012.

The first phixyn.com website probably went live
closer to 2010, but it was not Project Redhawk at
the time. Project Redhawk's official start date
was in fact 25th October 2012 and it aimed to be
a complete and modern re-design of the website,
adding in support for mobile web with a responsive
layout. CoolStoryBob