:wrench: add active_page to config array
Updates the example config file with the active_page
variable. This variable should be overriden before
including the header.php template, in order to
give the link in the top menu the 'is-active' class
and highlight it. This should only be done for pages
that have a top menu list entry, and the string
should match the comparison in the header.php file.