:sparkles: improve images used in blog post
Updates the blog post to wrap images with a figure element.
Adds links to the images so that the readers can click the
images to open the full image in a new tab.

This commit also includes an updated Spotify Wrapped 2019
image, which has been cropped nicely to remove the rough
edges and pink background :sparkles: