:hammer: make use of PHP vars in blog posts
Refactors the individual blog post pages to make
use of variables for blog post metadata. This
is fairly simple metadata for now, like date
posted, title and tag. When we do switch to
using a CMS, we can update the blog post
pages to retrieve more data from the CMS
database.