:package: add a simpler blog post template file
The example blog post holds a lot of examples of
things that can be used in future blog posts.
However, it can be annoying to use it as a
template and having to delete all of these
examples everytime we want to write a new post.
This commit adds a cut down blog post template
with just the things we need to get started. Until
we get a CMS up and running, this can be easily
re-used.