:package: make blog index page for final project
Adds a new tag to the blog for our UoS final project. Also adds
a blog post template for this tag with all the correct tag links
and name filled in. The tag index page contains 2 fake posts which
will need to be changed/removed prior to deploying to the server.