Core: Replace pages with a new content collection
Replace the existing pages. A new temporary collection for pages will be made
out of the new markdown files added here. As mentioned in the previous commit,
this is temporary until Strapi is set up and available for use. The idea is to
use this content to set up layouts and templates, and styles for the pages,
before we start pulling data from Strapi.