Core: Add a new template for the CustomPage collection
Adds a new template used to render content inside temp-content (i.e. the
CustomPage collection set up in the previous commits). For now it just
displays the content and sets the title. Note that because we want context
to access `this`, we need to make metaInfo() a function that returns the
metaInfo content.