Buefy: Start using our own scss file to customize Buefy
Add a new SCSS file to a folder called assets. Import this file in main.js.
Remove the old import for Buefy. I don't think it's needed (docs are
confusing) because we now have `@import "~buefy/src/scss/buefy";` in the
SCSS file, so it should be fine to remove it?