Components: Add a Carousel component
Add a carousel component that uses the Buefy Carousel. It's still WIP and
currently uses a data() function. It should use props instead, so that's
the next step. But it pretty much displays images and image previews as
indicators. The prev/next slide buttons are currently broke as we need to
add an icon pack.