Gridsome: Create pages for each project in the Strapi CMS
Use the Gridsome Pages API to create a new page for each project defined in
our Strapi CMS. We use GraphQL to query for pages, retrieving the slug, and
then loop through the results and call `createPage`.

This also adds the slug to the page's context, which can be accessed through
`$context` or inside `<page-query>` directly. See the code comments in the
modified file for more information.

This will make each page available at `/project/{slug}`. It will use the
'Project.vue' template for the page.