Components: Disable autoplay in Carousel
Add a prop for autoplay and set it to false by default. Autoplay can be
annoying if the carousel has a gallery. If the user clicks on an image
just as the interval reaches 0, the image will change and it can be
annoying. Autoplay can probably be set to true if the carousel has no
gallery (custom indicators), as that's more conventional.

Closes #75.