Pages: Wrap text around profile picture in about page for mobile
In the about page biography section (Card element), wrap text around the
profile picture on mobile. This gets rid of dead/white space and makes
text a bit easier to read for mobile users, as the lines are wider.

This change involved getting rid of the media component and replacing it
with columns. Sadly, it also made the profile picture display on the left
side on desktop, rather than right. I couldn't find a way to make it appear
on the right side with this solution.

Closes #74.