Fix: Fix footer not sticking to bottom of page
Fixes an issue wherein the footer would appear in the middle of the page if
the page didn't have enough content to fill the viewport vertically. This
fix involves making the app wrapper div a flex item and the section element
for the main content use flex-grow: 1. Have tested on desktop and mobile and
seems fine so far.