Components(Carousel): Add title and alt text to carousel images
Make the carousel image and carousel indicator images have an alt and title
text. Also change the figure element to use the fullwidth class. This prevents
the images from being squished to a particular ratio. Ideally we want all the
images in the carousel to be a particular ratio and for their ratio to be
passed in as a prop. A separate task has been made for this. For now, the
images will all appear in their individual aspect ratios, as I don't have time
to go through and fix each project's images.