Components(Header): Update navbar links and remove Buefy logo
Remove the placeholder Buefy logo from the navbar. Replace it with our name.
Add links to the social media items. Comment out the blog link for now.

This also adds the 'exact' property to the home path link. This makes it so
that this element does not get the active class set on every link. Really bad
at explaining this, so just see the router docs:
https://router.vuejs.org/api/#exact