Components(ProjectCard): Update card content to bind description html
Instead of having the project description inside a regular paragraph element,
use 'v-html' to bind its contents to a div. This will allow us to write HTML
for the project description and have it rendered in the card.

Also make the GitHub link open in a new tab, and improve the span elements
a little bit.