SEO: Add global/common meta tags
Update main.js to add common meta tags to the site's `<head>`. This is done
under the hood using Vue.meta.

This only adds tags that are common to every page. Tags like description and
preview images for social media cards will be added to each page individually.
The values for those will be stored in the CMS inside a meta_info component.

Docs:
https://gridsome.org/docs/head/
https://vue-meta.nuxtjs.org/