Gridsome: Move site version and start year to client config
So I discovered that global metadata can also go in the client config file,
and it looks much nicer if it's there with the rest of the config. So I've
moved it across and restored the original server config file with that
placeholder loadSource function. Cause you never know when you gonna need
that (probably never tbf).