Pages: Add meta tags to every page and project template
Add meta tags for Open Graph and Twitter card. Pages now query the global
metadata for site name and URL. Additionally, each page queries the CMS to
get its individual meta description and image values, which are used in
the og:description and og:image, and the Twitter counterparts.

The URL values are made up of the siteUrl and the $router.fullPath. This
ensures that it's 100% correct for every route. Also, everything is super
easy to change. The description and image can be changed from the CMS and
the URL can be changed from the config. Same with the site name.

Overall, I think it's a pretty decent way to have dynamic meta tags. Might
still be able to improve it though, but I'm pretty damn happy with it.