Pages(About): Move hobbies into a separate section
Add a new div on the About page for information about our hobbies, and remove
it from the biography card. This makes the biography slimmer and (hopefully)
easier to scan/read. The new hobbies section is similar to the "My projects"
section.

A new rich text field has been added to Strapi. As such, the GraphQL query
in this page has also been updated to pull this field.